backend/alert-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
slack-sdk==3.26.2
